chore: move test/contracts -> test-contracts, because mocha wouldn't otherwise run
problem is there's some ts in there and mocha throws up when it tries to process them for some reason
