+++ packages/smartcontracts/test-contracts/Oracle.d.ts
