+++ packages/smartcontracts/tsconfig.json
@@ -13,7 +13,7 @@
         "resolveJsonModule": true,
         "outDir": "dist"
     },
-    "include": ["./scripts", "./test", "./typechain"],
+    "include": ["./scripts", "./test", "./typechain", "test-contracts"],
     "files": [
       "./hardhat.config.ts"
     ]
